==> notebooks/01-Data-Storage.ipynb <==
import pandas as pd
import csv

from sqlalchemy import create_engine
from sqlalchemy import delete

from datetime import datetime

import sqlite3
import os

## Getting data from CSVs to DB

sample2020 = pd.read_csv('./VAERS-data/2020VAERSDATA/2020VAERSDATA.csv', encoding_errors = 'ignore', low_memory=False)

sample2020.head()

References:
https://github.com/kevinchiv/Predicting-Kickstarter-Success/blob/master/00%20-%20PostgreSQL%20Through%20Jupyter.ipynb
https://www.w3schools.com/sql/sql_datatypes.asp
https://datatofish.com/pandas-dataframe-to-sql/
https://datatofish.com/create-database-python-using-sqlite3/
https://dev.mysql.com/doc/connector-python/en/connector-python-api-mysqlcursor-execute.html
https://chartio.com/resources/tutorials/how-to-insert-if-row-does-not-exist-upsert-in-mysql/

def new_table_statement(df, table_name):
    
    column_names = df.columns.str.strip()
    datatypes = df.dtypes.values
    
    statement = 'CREATE TABLE IF NOT EXISTS ' + table_name + ' (\n'
    
    for idx, column_name in enumerate(column_names):
        dtype = datatypes[idx]
        statement += column_name.lower()
        if dtype == 'float64':
            statement += ' DECIMAL'
        elif dtype =='int64':
            statement += ' INTEGER'
        elif dtype == 'O':
            statement += ' VARCHAR'
        #denote primary key
        #if column_name == 'VAERS_ID':
            #statement += ' PRIMARY KEY'
        if idx != len(column_names) - 1:
            statement += ", \n"
    statement += ');'
    print(statement)
    return statement

new_table_statement(sample2020, 'test')

# https://stackoverflow.com/questions/9234560/find-all-csv-files-in-a-directory-using-python

def find_csvs(year):
    file_dir = os.path.abspath('.')
    data_folder = 'VAERS-data'
    data_year_folder = str(year) + 'VAERSData' #usual format
    path = os.path.join(file_dir, data_folder, data_year_folder)
    print(path)
    csv_list = list(filter(lambda x: '.csv' in x, os.listdir(path)))
    return csv_list, path

find_csvs(2020)

def make_delete_statement(table_name, year):
    statement = 'DELETE FROM ' + table_name + ' WHERE year = ' + str(year) + ';'
    return statement

make_delete_statement('test', 2021)

string = '2020VAERSVAX.csv'

string = string.split('/')[-1].split('.')[0]
table_name = string[9:].lower() 
print(table_name)

#only current year gets updated 

def make_tables(connection, year):
    this_year = datetime.now().year 
    csv_list, path = find_csvs(year)
    c = conn.cursor()
    for csv_file in csv_list:
        csv_path = os.path.join(path, csv_file)
        df = pd.read_csv(csv_path, encoding_errors = 'ignore', low_memory=False) 
        #making table name based on csv name 
        string = csv_file.split('/')[-1].split('.')[0]
        table_name = string[9:].lower() 
        year = int(string[:4])
        df['YEAR'] = year
        
        c.execute(new_table_statement(df, table_name))
        conn.commit()
        
        if this_year == year:
            #remove the rows with that year and replace with this updated csv data
            delete_statement = make_delete_statement(table_name, year)
            c.execute(delete_statement)
            conn.commit()
            print('Deleted any current data from given year!')
            #then replace
            df.to_sql(table_name, conn, if_exists = 'append', index=False)
        else:
            df.to_sql(table_name, conn, if_exists = 'append', index=False)
        conn.commit()

        print('Done! Inserted ' + table_name)
    

### Adding tables

conn = sqlite3.connect('vaers.db')

make_tables(conn, 2020)

make_tables(conn, 2021)

make_tables(conn, 2022)

#closing connection

c.close()
    
conn.close()
