==> notebooks/02-Data-Cleaning.ipynb <==
## From Database back into df (querying, etc.)


https://docs.sqlalchemy.org/en/14/tutorial/data.html

from sqlalchemy import create_engine
import pandas as pd

engine = create_engine("sqlite:///vaers.db")
engine

all_tables = engine.table_names()
all_tables

### SQL statements
VAERS data guide: https://vaers.hhs.gov/docs/VAERSDataUseGuide_en_September2021.pdf

Limiting to data for Covid-19 vaccines
- 1) Dataframe with: vaccine manufacturer, dose series number, vaccination date, number of days (calculated interval (in days) from the vaccination date to the onset date), age of patient (in years), and the Vaers ID for report
- 2) Dataframe with: symptoms, vaccine manufacturer, dose series number, year (of report)

Will lightly clean (remove nulls, group categories) before saving back to SQLite

#first query

s1 = '''
SELECT v.vax_manu, v.vax_dose_series, d.v_adminby, d.vax_date, d.numdays, d.age_yrs, v.vaers_id
FROM vax v 
INNER JOIN data d 
ON v.vaers_id = d.vaers_id
WHERE v.vax_type = 'COVID19';
'''

df1 = pd.read_sql(s1, engine)

df1.info()

df1.head()

df1.info()

#dropping nulls
df1 = df1.dropna(subset = ['vax_dose_series', 'v_adminby', 'numdays', 'age_yrs', 'vax_date'])

df1.v_adminby.value_counts()

dict_location = {'PVT': 'Private', 'PHM': 'Pharmacy', 'MIL': 'Military', 'OTH': 'Other', 'PUB': 'Public', 'WRK': 'Workplace Clinic', 'UNK': 'Unknown', 'OTH': 'Other', 'SEN': 'Nursing Home', 'SCH': 'School'}

df1.replace({'v_adminby': dict_location}, inplace=True)

df1.vax_dose_series.value_counts()

def group_multdoses(dose):
    if dose in ['5','6','7+']:
        dose = '5+'
    return dose

df1.vax_dose_series = df1.vax_dose_series.apply(lambda x: group_multdoses(x))

df1.vax_dose_series.value_counts()

# df1.to_pickle('df1.pkl')

# query for side effects table for vaers_id that matches for covid-19 vax

s2 = '''

SELECT vaers_id, year,
GROUP_CONCAT(COALESCE(symptom1, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(symptom2, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(symptom3, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(symptom4, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(symptom5, ''), ', ') AS all_symptoms
FROM symptoms 
WHERE vaers_id IN
(SELECT vaers_id 
FROM vax WHERE vax_type = 'COVID19')
GROUP BY vaers_id;
'''

symptoms_df = pd.read_sql(s2, engine)

symptoms_df

symptoms_df.vaers_id.nunique()

symptoms_df.iloc[2].all_symptoms

# getting more detailed df with symptoms, year, vaccine, and dose number

s3 = '''

SELECT s.vaers_id, s.year, v.vax_manu, v.vax_dose_series,
GROUP_CONCAT(COALESCE(s.symptom1, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(s.symptom2, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(s.symptom3, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(s.symptom4, ''),', ') ||', '||
GROUP_CONCAT(COALESCE(s.symptom5, ''), ', ') AS all_symptoms
FROM symptoms s
INNER JOIN vax v ON
s.vaers_id = v.vaers_id
WHERE v.vax_type = 'COVID19'
GROUP BY s.vaers_id;
'''

df3 = pd.read_sql(s3, engine)

df3

df3.info()

#dropping nulls
df3 = df3.dropna(subset = ['vax_manu', 'vax_dose_series'])

df3.vax_dose_series = df3.vax_dose_series.apply(lambda x: group_multdoses(x))

df3.vax_dose_series.value_counts()

### Saving dataframes back to SQLite as tables 
Saving these semi-cleaned dataframes back to SQL as new tables. The rest of the cleaning will be done via Spark.

df1.to_sql('stats', con=engine, if_exists = 'replace', index=False)

df3.to_sql('ade', con=engine,if_exists = 'replace', index=False)

Saving to CSV files too, for analysis with Spark:

df1.to_csv('stats.csv')
df3.to_csv('ade.csv')

## First analyzing with Pandas

Sources:
https://www.shanelynn.ie/summarising-aggregation-and-grouping-data-in-python-pandas/
https://proinsias.github.io/til/Pandas-Named-Aggregation/

import numpy as np

# summary stats

summary_df1 = df1.groupby(['vax_manu', 'vax_dose_series', 'v_adminby']).agg(
    avg_onset_days = pd.NamedAgg(column='numdays', aggfunc='mean'),
    std_onset_days= pd.NamedAgg(column='numdays', aggfunc='std'),
    avg_age = pd.NamedAgg(column='age_yrs', aggfunc='mean'),
    std_age = pd.NamedAgg(column='age_yrs', aggfunc='std'),
    cases = pd.NamedAgg(column='vaers_id', aggfunc='count')).reset_index()


summary_df1.to_pickle('summary_df1.pkl')

summary_df1.vax_manu.unique()
