==> notebooks/03-Spark-Analysis.ipynb <==
## Using Spark

import pandas as pd

import os
import pyspark

from pyspark.sql.functions import mean, stddev, count

from pyspark.sql.functions import split, explode, col, lower

from pyspark.sql.functions import collect_list, concat_ws

from pyspark.sql import SparkSession

#path to environment
os.environ['PYSPARK_PYTHON'] = '/Users/zliao/opt/anaconda3/envs/metis/bin/python'

spark = pyspark.sql.SparkSession.builder.config('spark.driver.memory', '6g').getOrCreate()

### Summary data reporting:

df1 = spark.read.csv('./stats.csv',sep=",", inferSchema="true", header="true")

df1.show(10)

https://sparkbyexamples.com/pyspark/pyspark-groupby-explained-with-example/

stats_vis = df1.groupby(['vax_manu', 'vax_dose_series', 'v_adminby'])\
    .agg(mean('numdays').alias('avg_onset'),\
        stddev('numdays').alias('std_onset'),\
        mean('age_yrs').alias('avg_age'),\
        stddev('age_yrs').alias('std_age'),\
        count('vaers_id').alias('cases'))

stats_vis.show(5)

#examining number of rows
stats_vis.count() 

#save! back to pandas since smaller now

stats_vis_df = stats_vis.toPandas()

stats_vis_df.to_csv('stats_vis.csv')

### Adverse effects reporting:

ades_df = spark.read.csv('./ade.csv',sep=",", inferSchema="true", header="true")

#enforce schema

ades_df.show(10)

https://stackoverflow.com/questions/41788919/concatenating-string-by-rows-in-pyspark

ades_df2 = ades_df.select(['vax_manu', 'vax_dose_series', 'all_symptoms']).groupby(['vax_manu', 'vax_dose_series']).agg(concat_ws(',', collect_list(ades_df.all_symptoms)).alias('symptoms'))

ades_df2.show(5)

from pyspark.sql.functions import udf
import re

https://stackoverflow.com/questions/46897988/removing-comma-in-a-column-in-pyspark

comma_clean = udf(lambda x: re.sub(" ,","", x))
#clean more in sql

ades_df2 = ades_df2.withColumn('cleaned_symptoms', comma_clean('symptoms'))

ades_df2.show(5)

#getting counts for each symptom (within manufacturer and dose series group)
(ades_df2.withColumn('separate', explode(split(lower(col('cleaned_symptoms')), ','))).groupby('separate').count().show(10))

ades_df3 = ades_df2.withColumn('cleaned_symptoms', explode(split(lower(col('cleaned_symptoms')), ','))).groupby(['vax_manu', 'vax_dose_series','cleaned_symptoms']).agg(count('cleaned_symptoms').alias('symptom_count'))

ades_df3.show()

https://stackoverflow.com/questions/38397796/retrieve-top-n-in-each-group-of-a-dataframe-in-pyspark

from pyspark.sql.window import Window

from pyspark.sql.functions import col, row_number

column_list = ["vax_manu","vax_dose_series"]
windowDept = Window.partitionBy([col(x) for x in column_list]).orderBy(col('symptom_count').desc())

ades_df4 = ades_df3.withColumn("row", row_number().over(windowDept)).filter(col("row") <= 10)

ades_df4.show()

#checking number of rows 
ades_df4.count()

ades_df4.filter('row = 1').show()

#save! back to pandas since smaller now

top_ades_df = ades_df4.toPandas()


top_ades_df.to_csv('top_ades.csv')

Supplemental:

top_ade_df = spark.sql('''
SELECT vax_manu, vax_dose_series, cleaned_symptoms
FROM
(SELECT *, ROW_NUMBER() OVER (PARTITION BY cleaned_symptoms ORDER BY symptom_count DESC) AS n 
FROM ade_counts) AS x
WHERE n <= 10;
''')
